Refactor helpers into multiple classes
